remove requirement for gpio controlled vbus switch
This controls the PSU output control using
SCPI commands to quickly enable/disable the output.
Instead of using a GPIO controlled switch.
Reducing the hardware needed.

There is a timing issue as the PSU I am testing
with is not very fast to process commands,
so the vbus enable timeout timer will sometimes
expire before the power supply turns on the output.
This happens in psource, probably due to
APP_PSOURCE_EN_TIMER_PERIOD
APP_PSOURCE_EN_MONITOR_TIMER_PERIOD
which then calls back to
app_psrc_tmr_cbk
and checks vbus_is_present
which can take a bit longer to come up than the configured
timeout.

But the output will power on to vsafe after 1 or 2 tries.
And then negociate up to EPR voltages.

There may be some additional timeout that is getting triggered
in the DPM state machine before the psource enable timers
causing the device to turn off before the PSU comes on.

Signed-off-by: Kieran Levin <ktl@frame.work>