do not repeat commands to PSU
Do not repeat output voltage commands if we already have the voltage set
to the same value.
The PSU will block turning on the output for some time after setting the
output voltage.
This causes the output to not be enabled before the type-c state machine
times out waiting for the output voltage to go high

Signed-off-by: Kieran Levin <ktl@frame.work>